chore: organize autocomplete unit tests
