+++ packages/autocomplete/test/unit/commands/autocomplete/script.unit.test.ts
@@ -1,6 +1,6 @@
 import {expect, test} from '@oclif/test'
 
-const {default: runtest} = require('../../helpers/runtest')
+const {default: runtest} = require('../../../helpers/runtest')
 
 runtest('autocomplete:script', () => {
   test
